+++ kubejs/startup_scripts/gtceu/aaa Materials and Elements/aa_elements.js
@@ -5,7 +5,7 @@ GTCEuStartupEvents.registry('gtceu:element', event => {
         177, //Neutrons
         -1, //Half Life Seconds
         null, //Decay To
-        'SiO+', //Atomic Symbol
+        'SiO₂+', //Atomic Symbol
         false//Is Isotope
         ) 
     event.create('mithril', 28, 178, -1, null, 'Mi', false) 
@@ -16,12 +16,12 @@ GTCEuStartupEvents.registry('gtceu:element', event => {
     event.create('desh', 32, 182, -1, null, 'De', false)
     event.create('ostrum', 33, 183, -1, null, 'Os', false)
     event.create('calorite', 34, 184, -1, null, 'Ct', false)
-    event.create('draconium', 35, 185, -1, null, 'Dc', false)
-    event.create('awakened_draconium', 36, 186, -1, null, 'Aw', false)
+    event.create('draconium', 35, 185, -1, null, 'דְרָקוֹן', false)
+    event.create('awakened_draconium', 36, 186, -1, null, 'לִויָתָן', false)
     event.create('radiance', 37, 187, -1, null, 'Ra', false)
     event.create('shadow_steel', 38, 188, -1, null, 'Sh', false)
     event.create('void', 39, 189, -1, null, '🕀', false)
-    event.create('stellar_matter', 40, 190, -1, null, '♁', false)
-    event.create('infinity', 41, 191, -1, null, '⳩', false)
+    event.create('stellar_matter', 999, 999, -1, null, '♁', false)
+    event.create('infinity', 999, 999, -1, null, '⳩', false)
     event.create('ardite', 42, 183, -1, null, 'Ar', false)
 })